cleaned up warnings

+++ OpenGL framework/src/GameObject/Components/Mesh.cpp
@@ -7,7 +7,7 @@ Mesh::Mesh(const std::string& fileName)
 
 void Mesh::InitMesh(const IndexedModel& model)
 {
-	m_numIndices = model.indices.size();
+	m_numIndices = (int)model.indices.size();
 
 	glGenVertexArrays(1, &m_vertexArrayObject);
 	glBindVertexArray(m_vertexArrayObject);
